% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_HY.R
\name{BIC_HY}
\alias{BIC_HY}
\title{Bayesian Information Criterion (BIC)}
\usage{
BIC_HY(Phi, n.data, n.par)
}
\arguments{
\item{Phi}{objective function value}

\item{n.data}{number of measured data}

\item{n.par}{number of adjustable parameters}
}
\description{
Bayesian Information Criterion (Schwarz, 1978) for least square estimations.
}
\references{
Ye, M., P.D. Meyer, and S.P. Neuman (2008): On model selection criteria in multimodel analysis. Water Resources Research 44 (3) W03428, doi:10.1029/2008WR006803.

Schwarz, G. (1978): Estimating the dimension of a model. The Annals of Statistics 6 (2), 461–464. URL: http://dx. doi. org/10.1214/aos/1176344136.

Peters and Durner (2015): SHYPFIT 2.0 User's Manual.
}
