% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKS.r
\name{calcKS}
\alias{calcKS}
\title{Saturated hydraulic conductivity}
\usage{
calcKS(V, Tmeas, L, A, dP)
}
\arguments{
\item{V}{water volume passed sample [L³]}

\item{Tmeas}{duration of measurement [time]}

\item{L}{length of the sample [L]}

\item{A}{cross-sectional area [L²]}

\item{dP}{pressure difference between top and bottom of the sample during the measurement [L]}
}
\value{
hydraulic conductivity (ks) [L/time].
}
\description{
Calculates saturated hydraulic conductivity (ks) following Darcy's law
}
\details{
Keep units consistent, e.g: V = cm³, dP = cm, A = cm², L = cm, Tmeas = hour
}
