% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSHP.R
\docType{data}
\name{dataSHP}
\alias{dataSHP}
\title{Soil hydraulic property data}
\format{
An object of class \code{data.frame} with 331 rows and 3 columns.
}
\usage{
data(dataSHP)
}
\description{
Soil hydraulic property data including soil water contents (th), unsaturated hydraulic conductivities (Ku) and the corresponding suctions/pressure heads.
}
\details{
\describe{\item{Columns:}{}
\item{}{Ku: unsaturated hydraulic conductivity}
\item{}{th: volumetric water content}
\item{}{suc: suction}}
}
\keyword{datasets}
