% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSE.R
\name{NSE}
\alias{NSE}
\title{Nash-Sutcliffe efficiency (NSE)}
\usage{
NSE(obs, sim)
}
\arguments{
\item{obs}{measured values}

\item{sim}{predicted values}
}
\description{
Nash-Sutcliffe efficiency (NSE)
}
\references{
Nash, J. E., and J.V. Sutcliffe (1970): River flow forecasting through conceptual models. 1. a discussion of principles. Journal of Hydrology 10, 282–290.
}
