% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitshp.R
\name{plot.fitSHP}
\alias{plot.fitSHP}
\title{Plot fitSHP object}
\usage{
\method{plot}{fitSHP}(x, ...)
}
\arguments{
\item{x}{object of class fitSHP}

\item{...}{arguments for plot}
}
\description{
Creates plot of fitSHP object with measured and fitted SWC, KU or both depending on fitSHP object
}
