% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Root mean square error (RMSE)}
\usage{
RMSE(obs, sim, na.rm = FALSE)
}
\arguments{
\item{obs}{measured values}

\item{sim}{predicted values}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Calculate Root mean square error (RMSE)
}
