% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSEM.R
\docType{data}
\name{dataSEM}
\alias{dataSEM}
\title{Evaporation experiment data}
\format{
An object of class \code{dataSEM} (inherits from \code{data.frame}) with 332 rows and 4 columns.
}
\usage{
data(dataSEM)
}
\description{
Example data of an Evaporation experiment
}
\details{
\describe{\item{Columns:}{}
\item{}{ts: timestamp [hour]}\item{}{weight: total weight of soil sample [g]}
\item{}{tens.up: measurements of upper tensiometer [cm]}
\item{}{tens.low: measurements of lower tensiometer [cm]}}
}
\keyword{datasets}
