% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Root mean square error (RMSE)}
\usage{
RMSE(obs, sim)
}
\arguments{
\item{obs}{measured values}

\item{sim}{predicted values}
}
\description{
Calculate Root mean square error (RMSE)
}
